+++ server/src/main/kotlin/miwayomi/api/ApiRoutes.kt
@@ -86,5 +86,6 @@ fun Application.registerApi() {
     registerExtensionApi()
     registerSourcePrefsApi()
     registerFavoritesApi()
+    registerWatchApi()
     registerUpdateApi()
 }